+++ src/components/BlockRenderer/BlockRenderer.jsx
@@ -9,15 +9,17 @@ import withObjectBrowser from '@plone/volto/components/manage/Sidebar/ObjectBrow
  * @extends Component
  */
 function BlockRenderer(props) {
-  const { edit, type } = props;
+  const { edit, type, blocksConfig } = props;
 
   if (!type) {
     // We could have an empty block, although should be handled somewhere else
     return null;
   }
 
-  const EditBlock = config.blocks.blocksConfig[type].edit;
-  const ViewBlock = config.blocks.blocksConfig[type].view;
+  const EditBlock =
+    blocksConfig?.[type]?.edit || config.blocks.blocksConfig[type].edit;
+  const ViewBlock =
+    blocksConfig?.[type]?.view || config.blocks.blocksConfig[type].view;
 
   if (!edit) {
     return <ViewBlock {...props} detached onChangeBlock={() => {}} />;
@@ -34,6 +36,7 @@ BlockRenderer.propTypes = {
   block: PropTypes.string.isRequired,
   onChangeBlock: PropTypes.func,
   data: PropTypes.objectOf(PropTypes.any).isRequired,
+  blocksConfig: PropTypes.objectOf(PropTypes.any),
 };
 
 BlockRenderer.defaultProps = {